chore(#31): move security folder inside, disable standard /login form
