+++ coincollector/src/main/java/io/github/lstramke/coincollector/security/RestAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package io.github.lstramke.security;
+package io.github.lstramke.coincollector.security;
 
 import java.io.IOException;
