+++ coincollector/src/test/java/io/github/lstramke/coincollector/handler/GroupHandlerTest.java
@@ -37,9 +37,9 @@ import io.github.lstramke.coincollector.exceptions.euroCoinCollectionGroupExcept
 import io.github.lstramke.coincollector.exceptions.euroCoinCollectionGroupException.EuroCoinCollectionGroupUpdateException;
 import io.github.lstramke.coincollector.model.EuroCoinCollectionGroup;
 import io.github.lstramke.coincollector.model.DTOs.Requests.UpdateGroupRequest;
+import io.github.lstramke.coincollector.security.SecurityConfig;
 import io.github.lstramke.coincollector.services.EuroCoinCollectionGroupStorageService;
 import io.github.lstramke.coincollector.services.SessionManager;
-import io.github.lstramke.security.SecurityConfig;
 
 @WebMvcTest(value = GroupHandler.class)
 @Import(SecurityConfig.class)
