+++ coincollector/src/test/java/io/github/lstramke/coincollector/handler/LoginHandlerTest.java
@@ -24,9 +24,9 @@ import org.springframework.test.web.servlet.MockMvc;
 
 import io.github.lstramke.coincollector.exceptions.userExceptions.UserNotFoundException;
 import io.github.lstramke.coincollector.model.User;
+import io.github.lstramke.coincollector.security.SecurityConfig;
 import io.github.lstramke.coincollector.services.SessionManager;
 import io.github.lstramke.coincollector.services.UserStorageService;
-import io.github.lstramke.security.SecurityConfig;
 
 @WebMvcTest(value = LoginHandler.class)
 @Import(SecurityConfig.class)
